changes to yolo architecture
