+++ src/misc/extract_image_classification_dataset.sh
