--- src/model_builder.py
@@ -1,267 +1,0 @@
-import torch
-import torch.nn as nn
-from torchinfo import summary
-
-
-class TinyVGG(nn.Module):
-    def __init__(self, hidden_units: int, output_shape: int):
-        """
-        ~1.6M parameters
-        """
-        super(TinyVGG, self).__init__()
-        assert False, "should set padding to all conv layers with kernel size"
-        num_channels = hidden_units
-
-        self.block_1 = nn.Sequential(
-            nn.Conv2d(3, num_channels, kernel_size=3, padding=1),
-            nn.BatchNorm2d(num_channels),
-            nn.ReLU(),
-            nn.Conv2d(num_channels, num_channels, kernel_size=3, padding=1),
-            nn.BatchNorm2d(num_channels),
-            nn.ReLU(),
-            nn.MaxPool2d(kernel_size=2)
-        )
-
-        self.block_2 = nn.Sequential(
-            nn.Conv2d(num_channels, num_channels, kernel_size=3, padding=1),
-            nn.BatchNorm2d(num_channels),
-            nn.ReLU(),
-            nn.Conv2d(num_channels, num_channels, kernel_size=3, padding=1),
-            nn.BatchNorm2d(num_channels),
-            nn.ReLU(),
-            nn.MaxPool2d(kernel_size=2)
-        )
-
-        self.block_3 = nn.Sequential(
-            nn.Conv2d(num_channels, num_channels, kernel_size=3, padding=1),
-            nn.BatchNorm2d(num_channels),
-            nn.ReLU(),
-            nn.Conv2d(num_channels, num_channels, kernel_size=3, padding=1),
-            nn.BatchNorm2d(num_channels),
-            nn.ReLU(),
-            nn.MaxPool2d(kernel_size=2)
-        )
-
-        self.block_4 = nn.Sequential(
-            nn.Conv2d(num_channels, num_channels, kernel_size=3, padding=1),
-            nn.BatchNorm2d(num_channels),
-            nn.ReLU(),
-            nn.Conv2d(num_channels, num_channels, kernel_size=3, padding=1),
-            nn.BatchNorm2d(num_channels),
-            nn.ReLU(),
-            nn.MaxPool2d(kernel_size=2)
-        )
-
-        self.classifier = nn.Sequential(
-            nn.Flatten(),
-            nn.Linear(num_channels*6*6, hidden_units),
-            nn.ReLU(),
-            nn.Linear(hidden_units, output_shape)
-        )
-
-    def forward(self, x: torch.Tensor):
-        x = self.block_1(x)
-        x = self.block_2(x)
-        x = self.block_3(x)
-        x = self.block_4(x)
-        x = self.classifier(x)
-        return x
-
-
-class DeepConvNet(nn.Module):
-    def __init__(self, dropout):
-        """
-        13M parameters
-        """
-        super(DeepConvNet, self).__init__()
-        self.block_1 = nn.Sequential(
-            nn.Conv2d(3, 64, kernel_size=7, stride=1),  # originally stride 2
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.MaxPool2d(kernel_size=2, stride=2)
-        )
-
-        self.block_2 = nn.Sequential(
-            nn.Conv2d(64, 192, kernel_size=3),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.MaxPool2d(kernel_size=2, stride=2)
-        )
-
-        self.block_3 = nn.Sequential(
-            nn.Conv2d(192, 128, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(128, 256, kernel_size=3),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(256, 256, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(256, 512, kernel_size=3),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.MaxPool2d(kernel_size=2, stride=2)
-        )
-
-        self.block_4 = nn.Sequential(
-            nn.Conv2d(512, 256, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(256, 512, kernel_size=3),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(512, 256, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(256, 512, kernel_size=3),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(512, 256, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(256, 512, kernel_size=3),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(512, 256, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(256, 512, kernel_size=3),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(512, 512, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(512, 1024, kernel_size=3),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.MaxPool2d(kernel_size=2, stride=2)
-        )
-
-        self.block_5 = nn.Sequential(
-            nn.Conv2d(1024, 512, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(512, 1024, kernel_size=3),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(1024, 512, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(512, 1024, kernel_size=3),
-            nn.LeakyReLU(negative_slope=0.1)
-        )
-
-        # adaptive avg pool is used to allow for any input size
-        # and output a fixed size tensor
-        self.avg_pool = nn.AdaptiveAvgPool2d((1, 1))
-        self.fc = nn.Linear(1024, 1000)
-        self.dropout = nn.Dropout(p=dropout)
-
-    def forward(self, x):
-        x = self.block_1(x)
-        x = self.block_2(x)
-        x = self.block_3(x)
-        x = self.block_4(x)
-        x = self.block_5(x)
-        x = self.avg_pool(x)
-        x = x.view(x.size(0), -1)  # batch size x 1024
-        x = self.dropout(x)
-        x = self.fc(x)
-        return x
-
-
-class YOLONet(nn.Module):
-    def __init__(self, dropout):
-        """
-
-        """
-        super(YOLONet, self).__init__()
-        self.block_1 = nn.Sequential(
-            nn.Conv2d(3, 64, kernel_size=7, stride=2,
-                      padding=3),  # originally stride 2
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.MaxPool2d(kernel_size=2, stride=2)
-        )
-
-        self.block_2 = nn.Sequential(
-            nn.Conv2d(64, 192, kernel_size=3, padding=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.MaxPool2d(kernel_size=2, stride=2)
-        )
-
-        self.block_3 = nn.Sequential(
-            nn.Conv2d(192, 128, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(128, 256, kernel_size=3, padding=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(256, 256, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(256, 512, kernel_size=3, padding=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.MaxPool2d(kernel_size=2, stride=2)
-        )
-
-        self.block_4 = nn.Sequential(
-            nn.Conv2d(512, 256, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(256, 512, kernel_size=3, padding=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(512, 256, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(256, 512, kernel_size=3, padding=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(512, 256, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(256, 512, kernel_size=3, padding=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(512, 256, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(256, 512, kernel_size=3, padding=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(512, 512, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(512, 1024, kernel_size=3, padding=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.MaxPool2d(kernel_size=2, stride=2)
-        )
-
-        self.block_5 = nn.Sequential(
-            nn.Conv2d(1024, 512, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(512, 1024, kernel_size=3, padding=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(1024, 512, kernel_size=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(512, 1024, kernel_size=3, padding=1),
-            nn.LeakyReLU(negative_slope=0.1)
-        )
-
-        self.block_5_part_2 = nn.Sequential(
-            nn.Conv2d(1024, 1024, kernel_size=3, padding=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(1024, 1024, kernel_size=3, padding=1, stride=2),
-            nn.LeakyReLU(negative_slope=0.1)
-        )
-
-        self.block_6 = nn.Sequential(
-            nn.Conv2d(1024, 1024, kernel_size=3, padding=1),
-            nn.LeakyReLU(negative_slope=0.1),
-            nn.Conv2d(1024, 1024, kernel_size=3, padding=1),
-            nn.LeakyReLU(negative_slope=0.1),
-        )
-
-        self.fc1 = nn.Linear(1024*7*7, 4096)
-        self.fc2 = nn.Linear(4096, 1470)
-
-        # adaptive avg pool is used to allow for any input size
-        # and output a fixed size tensor
-        # self.avg_pool = nn.AdaptiveAvgPool2d((1, 1))
-
-        self.dropout = nn.Dropout(p=dropout)
-
-    def forward(self, x):
-        x = self.block_1(x)
-        x = self.block_2(x)
-        x = self.block_3(x)
-        x = self.block_4(x)
-        x = self.block_5(x)
-        x = self.block_5_part_2(x)
-        x = self.block_6(x)
-        # x = x.view(x.size(0), -1)
-        # x = self.fc1(x)
-        # x = self.dropout(x)
-        # x = self.fc2(x)
-        # x = x.view(-1, 7, 7, 30)
-        return x
-
-
-if __name__ == "__main__":
-    # model = TinyVGG(hidden_units=128, output_shape=3)
-
-    # print(summary(model))
-    # print('-------')
-    model = YOLONet(dropout=0)
-    # print(summary(model))
-    summary(model, (1, 3, 448, 448))
